Added daemon to d2. Removed FNAL from available sites.

+++ installDataDealer.sh
@@ -4,16 +4,48 @@
 #
 # B.Barrefors (Aug 13, 2014)
 # --------------------------------------------------------------------------------------------------
-# parse arguments
-#================
+# Reset in case getopts has been used previously in the shell
+OPTIND=1
+
+# Initialize our own variables:
+testing=0
+keep_cache=0
+
+while getopts "tc" opt; do
+    case "$opt" in
+    t)  testing=1
+        ;;
+    c)  keep_cache=1
+        ;;
+    esac
+done
+
+# Configuration parameters
+export INTELROCCS_USER=cmsprod
+export INTELROCCS_GROUP=zh
+
+if [ testing -eq 0 ]
+then
+	INSTALL_DIR=/usr/local/IntelROCCS/Install
+	DATA_DEALER_BUDGET="50000"
+else
+	INSTALL_DIR=/usr/local/IntelROCCS/Install-test
+	DATA_DEALER_BUDGET="1000"
+fi
+
 # get src and install paths
 #==========================
 
 # get path of installation script, source code, and install path
 INTELROCCS_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 DATA_DEALER_SRC=${INTELROCCS_DIR}/DataDealer/src
-INSTALL_DIR=${INTELROCCS_DIR}/Install
 DATA_DEALER_INSTALL=${INSTALL_DIR}/DataDealer
+DATA_DEALER_LOG=/local/cmsprod/IntelROCCS/DataDealer
+PHEDEX_CACHE=/usr/local/IntelROCCS/Cache/DataDealer/Phedex
+POP_DB_CACHE=/usr/local/IntelROCCS/Cache/DataDealer/PopDb
+RANKINGS_CACHE=/usr/local/IntelROCCS/Cache/DataDealer/Rankings
+
+chown ${INTELROCCS_USER}:${INTELROCCS_GROUP} -R $DATA_DEALER_LOG
 
 # copy the software
 #==================
@@ -29,53 +61,49 @@ else
 fi
 # copy all source files to install directory
 cp $DATA_DEALER_SRC/* $DATA_DEALER_INSTALL
+cp $DATA_DEALER_SRC/../data_dealerd $DATA_DEALER_INSTALL
 touch ${DATA_DEALER_INSTALL}/__init__.py
 touch ${INSTALL_DIR}/__init__.py
 
-
-# install api
-#============
-
-API_INSTALL=${INSTALL_DIR}/Api
-${INTELROCCS_DIR}/installApi.sh
 
 # create init scripts
 #====================
 
 # set up cache environment
 # phedex
-#PHEDEX_CACHE=/tmp/IntelROCCS-dev/Cache/Phedex # used for testing
-PHEDEX_CACHE=/tmp/IntelROCCS/Cache/Phedex
 if [ -d "$PHEDEX_CACHE" ]
 then
 	# make sure to remove completely the previous installed software
-	echo " Cleaning up phedex cache."
-	rm -rf $PHEDEX_CACHE/*
-else
-	# create file structur if it doesn't exist
-	mkdir -p $PHEDEX_CACHE
+	if [ keep_cache -eq 0 ]
+	then
+		echo " Cleaning up phedex cache."
+		rm -rf $PHEDEX_CACHE/*
+	else
+		echo " Phedex cache is not cleared"
+	fi
 fi
+# create file structur if it doesn't exist
+mkdir -p $PHEDEX_CACHE
 
 # pop db
-#POP_DB_CACHE=/tmp/IntelROCCS-dev/Cache/PopDb # used for testing
-POP_DB_CACHE=/tmp/IntelROCCS/Cache/PopDb
 if [ -d "$POP_DB_CACHE" ]
 then
 	# make sure to remove completely the previous installed software
-	echo " Cleaning up pop db cache."
-	rm -rf $POP_DB_CACHE/*
-else
-	# create file structur if it doesn't exist
-	mkdir -p $POP_DB_CACHE
+	if [ keep_cache -eq 0 ]
+	then
+		echo " Cleaning up pop db cache."
+		rm -rf $POP_DB_CACHE/*
+	else
+		echo " Pop DB cache is not cleared"
+	fi
 fi
+# create file structur if it doesn't exist
+mkdir -p $POP_DB_CACHE
 
 PYTHONPATH=${INSTALL_DIR}/Api
 DATA_DEALER_THRESHOLD="1" # not used right now
-#DATA_DEALER_BUDGET="1000" # set low when testing
-DATA_DEALER_BUDGET="50000"
+
 CACHE_DEADLINE="12"
-#RANKING_CACHE=/tmp/IntelROCCS-dev/Cache # used for testing
-RANKING_CACHE=/tmp/IntelROCCS/Cache
 
 INIT_FILE=$DATA_DEALER_INSTALL/init.py
 
@@ -87,8 +115,33 @@ echo "os.environ['DATA_DEALER_POP_DB_CACHE']='"${POP_DB_CACHE}"'" >> $INIT_FILE
 echo "os.environ['DATA_DEALER_CACHE_DEADLINE']='"${CACHE_DEADLINE}"'" >> $INIT_FILE
 echo "os.environ['DATA_DEALER_THRESHOLD']='"${DATA_DEALER_THRESHOLD}"'" >> $INIT_FILE
 echo "os.environ['DATA_DEALER_BUDGET']='"${DATA_DEALER_BUDGET}"'" >> $INIT_FILE
-echo "os.environ['DATA_DEALER_RANKING_CACHE']='"${RANKING_CACHE}"'" >> $INIT_FILE
+echo "os.environ['DATA_DEALER_RANKINGS_CACHE']='"${RANKINGS_CACHE}"'" >> $INIT_FILE
 
 chmod 755 $INIT_FILE
+
+if [ testing -eq 0 ]
+then
+	# install and start daemons
+	#==========================
+
+	# stop potentially existing server process
+	if [ -e "/etc/init.d/data_dealerd" ]
+	then
+	  /etc/init.d/data_dealerd status
+	  /etc/init.d/data_dealerd stop
+	fi
+
+	# copy Detox daemon
+	cp ${INTELROCCS_DIR}/DataDealer/sysv/data_dealerd /etc/init.d/
+
+	# start new server
+	/etc/init.d/data_dealerd status
+	/etc/init.d/data_dealerd start
+	sleep 2
+	/etc/init.d/data_dealerd status
+
+	# start on boot
+	chkconfig --level 345 data_dealerd on
+fi
 
 exit 0
